mediacapture-streams/MediaStreamTrack-getCapabilities.https.html

<!doctype html>
<title>MediaStreamTrack GetCapabilities</title>
<p class="instructions">This test checks for the presence of audio and video properties in
<code>MediaStreamTrack.getCapabilities()</code> method.</p>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
  promise_test(() => {
  return navigator.mediaDevices.getUserMedia({audio: true})
    .then(stream => {
      var supported_constraints = navigator.mediaDevices.getSupportedConstraints();
      var capabilities = stream.getAudioTracks()[0].getCapabilities();
      if (supported_constraints.deviceId)
        assert_true(undefined !== capabilities.deviceId, "MediaTrackCapabilities's deviceId should exist.");
      if (supported_constraints.echoCancellation)
        assert_true(undefined !== capabilities.echoCancellation, "MediaTrackCapabilities's echoCancellation should exist.");
    });
  });

  promise_test(() => {
  return navigator.mediaDevices.getUserMedia({video: true})
    .then(stream => {
      var supported_constraints = navigator.mediaDevices.getSupportedConstraints();
      var capabilities = stream.getVideoTracks()[0].getCapabilities();
      if (supported_constraints.deviceId)
        assert_true(undefined !== capabilities.deviceId, "MediaTrackCapabilities's deviceId should exist.");
      if (supported_constraints.width)
        assert_true(undefined !== capabilities.width, "MediaTrackCapabilities's width should exist.");
      if (supported_constraints.height)
        assert_true(undefined !== capabilities.height, "MediaTrackCapabilities's height should exist.");
      if (supported_constraints.aspectRatio)
        assert_true(undefined !== capabilities.aspectRatio, "MediaTrackCapabilities's aspectRatio should exist.");
      if (supported_constraints.frameRate)
        assert_true(undefined !== capabilities.frameRate, "MediaTrackCapabilities's frameRate should exist.");
      if (supported_constraints.facingMode)
        assert_true(undefined !== capabilities.facingMode, "MediaTrackCapabilities's facingMode should exist.");
    });
  });
</script>
